Refactor ColorMixer component by moving it to a subdirectory and updating import paths in App component. Adjust layout for better responsiveness.
